1.Update the graphics' codes; 2. Update R.h and R.cpp.

+++ Sources/Elastos/Frameworks/Droid/Base/Core/inc/elastos/droid/graphics/CPaintFlagsDrawFilter.h
@@ -31,7 +31,7 @@ public:
         /* [in] */ Int32 setBits);
 
 private:
-    static CARAPI_(Int32) NativeConstructor(
+    static CARAPI_(Int64) NativeConstructor(
         /* [in] */ Int32 clearBits,
         /* [in] */ Int32 setBits);
